Remove superfluous i18n code.
Plugins on WordPress.org don't need to manage the POT file since it can be
generated automatically. This also allows the plugin to leverage the language
packs that have been available for plugins on WordPress.org for a few years.

This could potentially cause sites using custom language files to revert to
English. The benefits of the language packs and reduced maintenance far outweigh
any costs.